Fix bug in the Gelzinis and Valkunas MFGS code
The previous version had a bug where the actual coupling strength of the
GV MFGS approximation calculated was half of the value the user passed.
This was due to incorrect usage of the rescaled unit free temperature
and coupling strengths.

The new version, besided correcting this mistake, further generalizes
the expressions to be valid for arbitrary coupling angle.