Improve handling of principal value integrals for QWK expressions
The previous way of calculating the integrals seemed to depend strongly
on the epsilon interval around the singularity chosen. In this commit a
new method for computing the Cauchy principal value and Hadamard finite
part are introduced which should be much more accurate. With that said,
despite the integrals depending on the epsilon, the results of the QWK
calculation never seemed to depend on the epsilon, which may point to
the fact that the QWK expressions do not depend on the integrals
themeselves but to some combination of them that produces cancellations
of the errors; this however need further study to properly understand.

CAUTION: This current version fails to integrate the Delta and Delta'
integrals due to a numerical issue in the omega=0 boundary due to the
coth function. The full integrand is finite at oemga=0 (at least for
most spectral densities), but the current definition of functions
results in a NaN. Therefore, this version should not be used for
production until this issue is fixed.