Add new SpectralDensity method and fix QWK integrals
This commit adds a new method to spectral densities, 'sdinvω', which
returns the spectral density divided by omega. When defining a new
spectral density type, either 'sd' or 'sdinvω' must be provided.
The new method is particularly useful for several of the spectral
densities integrals one has to compute. In particular, this allows for a
relatively straightfoward fix for the issue of evaluating the integrand
of the Delta and Delta' integrals at the boundary values.